services/trip/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic<2
python-jose[cryptography] 
passlib[bcrypt]
aiokafka
python-dotenv